[FIX] sale_order_payment: fix product_barcodes
