+++ additional_product_imporpalac/models/product_product.py
@@ -14,20 +14,3 @@ class ProductProduct(models.Model):
             record_write_date = record.write_date or self.env.cr.now()
             tmpl_write_date = record.product_tmpl_id.write_date or datetime.min
             record.write_date = max(record_write_date, tmpl_write_date)
-
-
-class ProductTemplate(models.Model):
-    _inherit = "product.template"
-
-    @api.depends("product_variant_ids.barcode_supp")
-    def _compute_barcodes(self):
-        res = super()._compute_barcodes()
-        for template in self:
-            current_barcodes = (
-                template.barcodes if isinstance(template.barcodes, list) else []
-            )
-            new_barcodes = [
-                p.barcode_supp for p in template.product_variant_ids if p.barcode_supp
-            ]
-            template.barcodes = list(dict.fromkeys(current_barcodes + new_barcodes))
-        return res
