+++ stock_picking_flow/models/__init__.py
@@ -1,4 +1,3 @@
 from . import stock_picking
 from . import stock_move
-from . import product_template
 from . import purchase_order
