+++ supplier_barcode/models/product_barcode_generator.py
@@ -71,6 +71,19 @@ class ProductTemplate(models.Model):
                 rec.barcode_supp = ean
                 rec.generate_barcode_image_supp(ean)
 
+    @api.depends("product_variant_ids.barcode_supp")
+    def _compute_barcodes(self):
+        res = super()._compute_barcodes()
+        for template in self:
+            current_barcodes = (
+                template.barcodes if isinstance(template.barcodes, list) else []
+            )
+            new_barcodes = [
+                p.barcode_supp for p in template.product_variant_ids if p.barcode_supp
+            ]
+            template.barcodes = list(dict.fromkeys(current_barcodes + new_barcodes))
+        return res
+
 
 class ShProductProduct(models.Model):
     _inherit = "product.product"
